+++ src/ilcdlib/mapping/common.py
@@ -23,7 +23,7 @@ __all__ = (
 )
 
 import abc
-from typing import Generic, TypeVar
+from typing import Generic, TypeVar, cast
 
 T = TypeVar("T")
 K = TypeVar("K")
@@ -56,3 +56,22 @@ class SimpleDataMapper(BaseDataMapper[T, T], Generic[T]):
         :param default_value: The default value to return if there is no mapping for input value.
         """
         return self.DATABASE.get(input_value, default_value)
+
+
+class KeyValueMapper(BaseDataMapper[str, T], Generic[T]):
+    """A data mapper that maps input values to output values using keywords."""
+
+    KV: dict[str, list[T]] = {}
+
+    def map(self, input_value: str, default_value: T | None) -> T | None:
+        """
+        Map the input value to the output value using keywords.
+
+        :param input_value: The input value to map.
+        :param default_value: The default value to return if there is no mapping for input value.
+        """
+        for impact_name, keywords in self.KV.items():
+            for keyword in keywords:
+                if str(keyword).strip().lower() in input_value.strip().lower():
+                    return cast(T, impact_name)
+        return default_value