+++ src/ilcdlib/mapping/impacts.py
@@ -17,7 +17,8 @@
 #  Charles Pankow Foundation, Microsoft Sustainability Fund, Interface, MKA Foundation, and others.
 #  Find out more at www.BuildingTransparency.org
 #
-from ilcdlib.mapping.common import SimpleDataMapper
+from ilcdlib.mapping.common import BaseDataMapper, K, KeyValueMapper, SimpleDataMapper
+from ilcdlib.utils import is_valid_uuid
 
 
 class ImpactsUUIDToOpenIdMapper(SimpleDataMapper[str]):
@@ -41,4 +42,37 @@ class ImpactsUUIDToOpenIdMapper(SimpleDataMapper[str]):
     }
 
 
-default_impacts_uuid_mapper = ImpactsUUIDToOpenIdMapper()
+class ImpactsKeywordToOpenIdMapper(KeyValueMapper[str]):
+    """Map keywords to openEPD impact names.""" ""
+
+    KV = {
+        "gwp-biogenic": ["biogenic"],
+        "gwp-luluc": ["luluc"],
+        "gwp-fossil": ["fossil"],
+        "gwp-gwp_nonCO2": ["CO2"],
+        "gwp": ["gwp", "global warming"],
+        "ep-marine": ["marine"],
+        "ep-fresh": ["fresh", "fw"],
+        "ep-terr": ["terr"],
+        "ep": ["ep", "eutrophication"],
+        "odp": ["odp", "depletion"],
+        "ap": ["ap", "acidification"],
+        "pocp": ["pocp", "photochemical", "smog", "ozone creation"],
+    }
+
+
+class DefaultImpactsToOpenIdMapper(BaseDataMapper[str, str]):
+    """Map default impacts to openEPD impact names."""
+
+    def __init__(self):
+        self._uuid_mapper = ImpactsUUIDToOpenIdMapper()
+        self._keyword_mapper = ImpactsKeywordToOpenIdMapper()
+
+    def map(self, input_value: str, default_value: str | None) -> K | None:
+        """Map the input value to the output value."""
+        if is_valid_uuid(input_value):
+            return self._uuid_mapper.map(input_value, default_value)
+        return self._keyword_mapper.map(input_value, default_value)
+
+
+default_impacts_mapper = DefaultImpactsToOpenIdMapper()